refactor: 기본 규칙 섹션 분리
