+++ apps/admin/src/app/(admin)/measurement-rules/_components/measurement-rule-form/measurement-rule-form.tsx
@@ -12,6 +12,8 @@ import {
   useWatch,
 } from "react-hook-form";
 
+import { MeasurementRuleDefaultSection } from "./default-section";
+
 import { BasicAlert } from "@/components/Alert";
 import { Button } from "@/components/ui/button";
 import {
@@ -69,7 +71,6 @@ interface MeasurementRuleFormData extends MeasurementRule {
   sleeveType?: SleeveType;
   necklineType?: NecklineType;
   name: string;
-  duplicateError: boolean;
 }
 
 export function MeasurementRuleForm({
@@ -83,7 +84,6 @@ export function MeasurementRuleForm({
       categoryId: "",
       name: "",
       items: [],
-      duplicateError: false, // TODO: 중복 체크 오류 표시 위해 추가
       level1: categories[0]?.id || "",
       level2: categories[0]?.children?.[0]?.id || "",
     },
@@ -153,7 +153,10 @@ export function MeasurementRuleForm({
       requestData.sleeveType
     );
     if (isDuplicate) {
-      form.setValue("duplicateError", true);
+      form.setError("name", {
+        message: "중복된 규칙입니다.",
+      });
+      // form.setValue("duplicateError", true);
       // setDuplicateError(true);
       return;
     }
@@ -177,27 +180,7 @@ export function MeasurementRuleForm({
             </CardDescription>
           </CardHeader>
           <CardContent className="space-y-6">
-            {/* 에러 메시지 표시 */}
-            {form.getValues().duplicateError && (
-              <BasicAlert title="중복 오류" variant="destructive">
-                선택한 카테고리와 소매 유형의 조합으로 이미 치수 규칙이
-                존재합니다. 다른 조합을 선택해주세요.
-              </BasicAlert>
-            )}
-
-            {/* 카테고리 선택 */}
-            <div className="space-y-4">
-              <div>
-                <FormLabel>카테고리 선택</FormLabel>
-                <FormDescription>
-                  대분류, 중분류, 소분류를 순서대로 선택해주세요.
-                </FormDescription>
-              </div>
-              <CategorySelect />
-            </div>
-
-            <TopNeedFieldForm />
-            <MeasurementRuleName />
+            <MeasurementRuleDefaultSection />
           </CardContent>
         </Card>
 
@@ -266,94 +249,6 @@ export function MeasurementRuleForm({
       </form>
       <QueryDevTools control={form.control} />
     </Form>
-  );
-}
-
-function FormSelect({
-  options,
-  label,
-  placeholder,
-  onChange,
-  ...props
-}: {
-  options: { id: string; name: string }[];
-  label?: string;
-  placeholder?: string;
-  name: ComponentProps<typeof FormField>["name"];
-  onChange?: (value: string) => void;
-}) {
-  return (
-    <FormField
-      {...props}
-      render={({ field }) => (
-        <FormItem>
-          {label && <FormLabel>{label}</FormLabel>}
-          <FormControl>
-            <Select
-              value={field.value}
-              onValueChange={(value) => {
-                field.onChange(value);
-                onChange?.(value);
-              }}
-            >
-              <SelectTrigger>
-                <SelectValue placeholder={placeholder} />
-              </SelectTrigger>
-              <SelectContent>
-                {options.map((option) => (
-                  <SelectItem key={option.id} value={option.id.toString()}>
-                    {option.name}
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-          </FormControl>
-        </FormItem>
-      )}
-    />
-  );
-}
-
-function CategorySelect() {
-  const form = useFormContext();
-
-  const level1 = useWatch({ name: "level1" });
-  const level2 = useWatch({ name: "level2" });
-
-  // 대-중-소 카테고리 리스트
-  const level1Categories = categories.filter((cat) => cat.parent_id === null);
-  const level2Categories =
-    level1Categories.find((cat) => cat.id == level1)?.children || [];
-  const level3Categories =
-    level2Categories.find((cat) => cat.id === level2)?.children || [];
-
-  return (
-    <div className="grid grid-cols-3 gap-4">
-      <FormSelect
-        name="level1"
-        options={level1Categories}
-        placeholder="대분류"
-        onChange={() => {
-          form.setValue("level2", null);
-          form.setValue("level3", null);
-        }}
-      />
-      <FormSelect
-        name="level2"
-        options={level2Categories}
-        placeholder="중분류"
-        onChange={() => form.setValue("level3", null)}
-      />
-      <FormSelect
-        name="level3"
-        options={level3Categories}
-        placeholder="소분류"
-        onChange={(value) => {
-          form.setValue("categoryId", value);
-          form.clearErrors("duplicateError");
-        }}
-      />
-    </div>
   );
 }
 
@@ -525,92 +420,6 @@ function RuleCheckItem({
         </label>
       </div>
     </div>
-  );
-}
-
-/**
- * 자동 생성되는 규칙 이름 표시
- */
-function MeasurementRuleName() {
-  const form = useFormContext();
-
-  const categoryLevel3 = useWatch({ name: "level3" });
-  const category = getCategoryById(categoryLevel3);
-
-  const categoryLevel2 = useWatch({ name: "level2" });
-  const selectedSleeveType = useWatch({ name: "sleeveType" });
-  const selectedNecklineType = useWatch({ name: "necklineType" });
-
-  const getAuthName = () => {
-    // 자동 생성되며 수동 수정 불가
-    const category3Name = category?.name ?? "";
-    switch (categoryLevel2) {
-      // - 상의: `넥라인 + 소매 유형 + 소분류`
-      case CATEGORY_ID.상의: {
-        return `${selectedNecklineType} ${selectedSleeveType} ${category3Name}`;
-      }
-      // - 그 외: `소분류명`
-      default:
-        return category3Name;
-    }
-  };
-
-  const authName = getAuthName();
-
-  useEffect(() => {
-    form.setValue("name", authName);
-  }, [authName]);
-
-  return (
-    <FormField
-      name="name"
-      render={({ field }) => (
-        <FormItem>
-          <FormLabel>규칙 이름</FormLabel>
-          <FormControl>
-            <Input {...field} readOnly placeholder="자동으로 생성됩니다" />
-          </FormControl>
-          <FormDescription>
-            소매 유형과 카테고리를 선택하면 자동으로 설정됩니다.
-          </FormDescription>
-          <FormMessage />
-        </FormItem>
-      )}
-    />
-  );
-}
-
-function TopNeedFieldForm() {
-  const form = useFormContext();
-  const categoryLevel2 = useWatch({ name: "level2" });
-
-  if (categoryLevel2 !== CATEGORY_ID.상의) return null;
-
-  return (
-    <>
-      <FormSelect
-        label="소매 유형"
-        name="sleeveType"
-        options={SLEEVE_TYPES.map((type) => ({
-          id: type,
-          name: type,
-        }))}
-        onChange={() => {
-          form.clearErrors("duplicateError");
-        }}
-      />
-      <FormSelect
-        label="넥라인"
-        name="necklineType"
-        options={NECKLINE_TYPES.map((type) => ({
-          id: type,
-          name: type,
-        }))}
-        onChange={() => {
-          form.clearErrors("duplicateError");
-        }}
-      />
-    </>
   );
 }
