+++ apps/admin/src/queries/measurement-rule.ts
@@ -3,6 +3,8 @@ import { queryOptions } from "@tanstack/react-query";
 import {
   getMeasurementRuleItemCode,
   GetMeasurementRuleItemCodeResponse,
+  getMeasurementRuleList,
+  GetMeasurementRuleListResponse,
 } from "@/services/measurement-rule";
 
 const getMeasurementRuleItemCodeQueryOptions = () => {
@@ -12,6 +14,14 @@ const getMeasurementRuleItemCodeQueryOptions = () => {
   });
 };
 
+const getMeasurementRuleListQueryOptions = () => {
+  return queryOptions<GetMeasurementRuleListResponse>({
+    queryKey: ["measurement-rule", "list"],
+    queryFn: () => getMeasurementRuleList(),
+  });
+};
+
 export const measurementRuleQueries = {
   getMeasurementRuleItemCodeQueryOptions,
+  getMeasurementRuleListQueryOptions,
 };
